Convert the map hash to JSON and back to a hash
We need to do this as we are converting hashes and input keys
from JSON, this allows us to be certain the keys on the input
map are strings. The key that is sent to the map function can
then be converted to a string before using it to dig into the
hash.

Co-authored-by: Jaseera <jaseera@madetech.com>